WIP: Added test config and other minor fixes
